+++ database/seeders/DatabaseSeeder.php
@@ -16,7 +16,6 @@ class DatabaseSeeder extends Seeder
         $this->call([
             RolesAndPermissionsSeeder::class,
             AdminUserSeeder::class,
-            SystemSettingsSeeder::class,
             MenuSeeder::class,
         ]);
     }