+++ database/settings/2025_11_14_233329_create_maintenance_settings.php
@@ -1,0 +1,13 @@
+<?php
+
+use Spatie\LaravelSettings\Migrations\SettingsMigration;
+
+return new class extends SettingsMigration
+{
+    public function up(): void
+    {
+        $this->migrator->add('maintenance.maintenance_mode', false);
+        $this->migrator->add('maintenance.maintenance_message', 'We are currently performing scheduled maintenance. We will be back shortly.');
+        $this->migrator->add('maintenance.maintenance_end', null);
+    }
+};